device-flasher: Add support for upgrading devices with OTAs
The current flasher is only capable of flashing entire factory
images, which unlock/lock the device, and wipes userdata in the
process.

After this commit you can pass an '--ota' argument to CLI to opt
to flash an OTA package in the current directory. This is useful
for upgrading devices without needing to reprovision.

Also make string matching a bit better in getPrerequisiteFiles().

Change-Id: Ib85e8802056be00c31b06c1e9110a16b5c1dd9fb